modify

+++ include/types.h
@@ -1,6 +1,6 @@
 #ifndef TYPES_H
 #define TYPES_H
 
-using scalar = double;
+using scalar = float;
 
 #endif
